initial removal of pre-phase1 code
