+++ RecoLocalCalo/HcalRecAlgos/interface/HcalDeterministicFit.h
@@ -1,5 +1,5 @@
-#ifndef HcalDeterministicFit_h
-#define HcalDeterministicFit_h 1
+#ifndef RecoLocalCalo_HcalRecAlgos_HcalDeterministicFit_h
+#define RecoLocalCalo_HcalRecAlgos_HcalDeterministicFit_h
 
 #include <typeinfo>
 #include <vector>
@@ -25,8 +25,6 @@ class HcalDeterministicFit {
 		   float& reconstructedTime) const;
 
   // This is the CMSSW Implementation of the apply function
-  template<class Digi>
-  void apply(const CaloSamples & cs, const std::vector<int> & capidvec, const HcalCalibrations & calibs, const Digi & digi, double& ampl, float &time) const;
   void getLandauFrac(float tStart, float tEnd, float &sum) const;
   void get205Frac(float tStart, float tEnd, float &sum) const;
 
@@ -77,106 +75,5 @@ class HcalDeterministicFit {
     0.01149, 0.0112412, 0.0110002 };
 
 };
-
-template<class Digi>
-void HcalDeterministicFit::apply(const CaloSamples & cs, const std::vector<int> & capidvec, const HcalCalibrations & calibs, const Digi & digi, double & reconstructedEnergy, float & reconstructedTime) const {
-  std::vector<double> corrCharge;
-  std::vector<double> inputCharge;
-  std::vector<double> inputPedestal;
-  double gainCorr = 0;
-  double respCorr = 0;
-
-  for(int ip=0; ip<cs.size(); ip++){
-    const int capid = capidvec[ip];
-    double charge = cs[ip];
-    double ped = calibs.pedestal(capid);
-    double gain = calibs.respcorrgain(capid);
-    gainCorr = gain;
-    inputCharge.push_back(charge);
-    inputPedestal.push_back(ped);
-  }
-
-  fPedestalSubFxn_.calculate(inputCharge, inputPedestal, corrCharge);
-
-  const HcalDetId& cell = digi.id();
-  double fpar0, fpar1, fpar2;
-  if(std::abs(cell.ieta())<HcalRegion[0]){
-    fpar0 = fpars[0];
-    fpar1 = fpars[1];
-    fpar2 = fpars[2];
-  }else if(std::abs(cell.ieta())==HcalRegion[0]||std::abs(cell.ieta())==HcalRegion[1]){
-    fpar0 = fpars[3];
-    fpar1 = fpars[4];
-    fpar2 = fpars[5];
-  }else{
-    fpar0 = fpars[6];
-    fpar1 = fpars[7];
-    fpar2 = fpars[8];
-  }
-
-  if (fTimeSlew==0)respCorr=1.0;
-  else if (fTimeSlew==1)respCorr=rCorr[0];
-  else if (fTimeSlew==2)respCorr=rCorr[1];
-  else if (fTimeSlew==3)respCorr=frespCorr;
-
-  float tsShift3=HcalTimeSlew::delay(inputCharge[3], fTimeSlew, fTimeSlewBias, fpar0, fpar1 ,fpar2);
-  float tsShift4=HcalTimeSlew::delay(inputCharge[4], fTimeSlew, fTimeSlewBias, fpar0, fpar1 ,fpar2);
-  float tsShift5=HcalTimeSlew::delay(inputCharge[5], fTimeSlew, fTimeSlewBias, fpar0, fpar1 ,fpar2);
-
-  float i3=0;
-  getLandauFrac(-tsShift3,-tsShift3+tsWidth,i3);
-  float n3=0;
-  getLandauFrac(-tsShift3+tsWidth,-tsShift3+tsWidth*2,n3);
-  float nn3=0;
-  getLandauFrac(-tsShift3+tsWidth*2,-tsShift3+tsWidth*3,nn3);
-
-  float i4=0;
-  getLandauFrac(-tsShift4,-tsShift4+tsWidth,i4);
-  float n4=0;
-  getLandauFrac(-tsShift4+tsWidth,-tsShift4+tsWidth*2,n4);
-
-  float i5=0;
-  getLandauFrac(-tsShift5,-tsShift5+tsWidth,i5);
-  float n5=0;
-  getLandauFrac(-tsShift5+tsWidth,-tsShift5+tsWidth*2,n5);
-
-  float ch3=0;
-  float ch4=0;
-  float ch5=0;
-
-  if (i3 != 0 && i4 != 0 && i5 != 0) {
-
-    ch3=corrCharge[3]/i3;
-    ch4=(i3*corrCharge[4]-n3*corrCharge[3])/(i3*i4);
-    ch5=(n3*n4*corrCharge[3]-i4*nn3*corrCharge[3]-i3*n4*corrCharge[4]+i3*i4*corrCharge[5])/(i3*i4*i5);
-
-    if (ch3<negThresh[0]) {
-      ch3=negThresh[0];
-      ch4=corrCharge[4]/i4;
-      ch5=(i4*corrCharge[5]-n4*corrCharge[4])/(i4*i5);
-    }
-    if (ch5<negThresh[0] && ch4>negThresh[1]) {
-      double ratio = (corrCharge[4]-ch3*i3)/(corrCharge[5]-negThresh[0]*i5);
-      if (ratio < 5 && ratio > 0.5) {
-        double invG = invGpar[0]+invGpar[1]*std::sqrt(2*std::log(invGpar[2]/ratio));
-        float iG=0;
-        getLandauFrac(-invG,-invG+tsWidth,iG);
-        if (iG != 0 ) {
-	  ch4=(corrCharge[4]-ch3*n3)/(iG);
-	  tsShift4=invG;
-	}
-      }
-    }
-  }
-
-  if (ch4<1) {
-    ch4=0;
-  }
-
-  double ampl=ch4*gainCorr*respCorr;
-  reconstructedEnergy=ampl;
-  reconstructedTime=tsShift4;
-}
-
 
 #endif // HLTAnalyzer_h
