+++ RecoLocalCalo/HcalRecAlgos/interface/PulseShapeFitOOTPileupCorrection.h
@@ -1,14 +1,13 @@
-#ifndef PulseShapeFitOOTPileupCorrection_h
-#define PulseShapeFitOOTPileupCorrection_h 1
+#ifndef RecoLocalCalo_HcalRecAlgos_PulseShapeFitOOTPileupCorrection_h
+#define RecoLocalCalo_HcalRecAlgos_PulseShapeFitOOTPileupCorrection_h
 
 #include <typeinfo>
 
-#include "DataFormats/HcalDetId/interface/HcalDetId.h"
+#include "RecoLocalCalo/HcalRecAlgos/interface/PulseShapeFunctor.h"
+
 #include "DataFormats/HcalRecHit/interface/HBHEChannelInfo.h"
 #include "CalibCalorimetry/HcalAlgos/interface/HcalPulseShapes.h"
 #include "CalibCalorimetry/HcalAlgos/interface/HcalTimeSlew.h"
-#include "CalibFormats/HcalObjects/interface/HcalCoder.h"
-#include "CalibFormats/HcalObjects/interface/HcalCalibrations.h"
 
 #include <TMinuit.h>
 
@@ -19,76 +18,6 @@
 
 #include "RecoLocalCalo/HcalRecAlgos/src/HybridMinimizer.h"
 
-namespace HcalConst{
-
-   constexpr int maxSamples = 10;
-   constexpr int maxPSshapeBin = 256;
-   constexpr int nsPerBX = 25;
-   constexpr float iniTimeShift = 92.5f;
-   constexpr double invertnsPerBx = 0.04;
-   constexpr int shiftTS = 4;
-
-}
-
-namespace FitterFuncs{
-  
-   class PulseShapeFunctor {
-      public:
-     PulseShapeFunctor(const HcalPulseShapes::Shape& pulse,bool iPedestalConstraint, bool iTimeConstraint,bool iAddPulseJitter,bool iAddTimeSlew,
-		       double iPulseJitter,double iTimeMean,double iTimeSig,double iPedMean,double iPedSig,
-		       double iNoise, unsigned int nSamplesToFit);
-     ~PulseShapeFunctor();
-     
-     double EvalPulse(const double *pars, unsigned int nPar);
-     
-     void setDefaultcntNANinfit(){ cntNANinfit =0; }
-     int getcntNANinfit(){ return cntNANinfit; }
-     
-     void setpsFitx(double *x ){ for(int i=0; i<HcalConst::maxSamples; ++i) psFit_x[i] = x[i]; }
-     void setpsFity(double *y ){ for(int i=0; i<HcalConst::maxSamples; ++i) psFit_y[i] = y[i]; }
-     void setpsFiterry (double *erry  ){ for(int i=0; i<HcalConst::maxSamples; ++i) psFit_erry  [i] = erry [i]; }
-     void setpsFiterry2(double *erry2 ){ for(int i=0; i<HcalConst::maxSamples; ++i) psFit_erry2 [i] = erry2[i]; }
-     void setpsFitslew (double *slew  ){ for(int i=0; i<HcalConst::maxSamples; ++i) {psFit_slew [i] = slew [i]; } }
-     double sigmaHPDQIE8(double ifC);
-     double sigmaSiPMQIE10(double ifC);
-     double getSiPMDarkCurrent(double darkCurrent, double fcByPE, double lambda);
-     void setinvertpedSig2(double x) { invertpedSig2_ = x; }
-
-     double singlePulseShapeFunc( const double *x );
-     double doublePulseShapeFunc( const double *x );
-     double triplePulseShapeFunc( const double *x );
-     
-   private:
-     std::array<float,HcalConst::maxPSshapeBin> pulse_hist;
-     
-     int cntNANinfit;
-     std::vector<float> acc25nsVec, diff25nsItvlVec;
-     std::vector<float> accVarLenIdxZEROVec, diffVarItvlIdxZEROVec;
-     std::vector<float> accVarLenIdxMinusOneVec, diffVarItvlIdxMinusOneVec;
-     void funcHPDShape(std::array<double,HcalConst::maxSamples> & ntmpbin, const double &pulseTime, const double &pulseHeight,const double &slew);
-     double psFit_x[HcalConst::maxSamples], psFit_y[HcalConst::maxSamples], psFit_erry[HcalConst::maxSamples], psFit_erry2[HcalConst::maxSamples], psFit_slew[HcalConst::maxSamples];
-     
-     unsigned nSamplesToFit_;
-     bool pedestalConstraint_;
-     bool timeConstraint_;
-     bool addPulseJitter_;
-     bool unConstrainedFit_;
-     double pulseJitter_;
-     double timeMean_;
-     double timeSig_;
-     double pedMean_;
-     double pedSig_;
-     double noise_;
-     double timeShift_;
-
-     double inverttimeSig_, inverttimeSig2_;
-     double invertpedSig_, invertpedSig2_;
-     std::array<double,HcalConst::maxSamples> pulse_shape_;
-     std::array<double,HcalConst::maxSamples> pulse_shape_sum_;
-
-   };
-   
-}
 
 class PulseShapeFitOOTPileupCorrection
 {
@@ -102,20 +31,11 @@ public:
 		     bool & useTriple,
 		     float& chi2) const;
 
-    void apply(const CaloSamples & cs,
-	       const std::vector<int> & capidvec,
-	       const HcalCalibrations & calibs,
-	       double& reconstructedEnergy,
-	       float& reconstructedTime,
-	       bool & useTriple,
-	       float& chi2) const;
-
     void setPUParams(bool   iPedestalConstraint, bool iTimeConstraint,bool iAddPulseJitter,bool iApplyTimeSlew,
 		     double iTS4Min, const std::vector<double> & iTS4Max,
 		     double iPulseJitter,
 		     double iTimeMean, double iTimeSigHPD, double iTimeSigSiPM,
-		     double iPedMean, double iPedSigHPD, double iPedSigSiPM,
-		     double iNoiseHPD, double iNoiseSiPM,
+		     double iPedMean,
 		     double iTMin, double iTMax,
 		     const std::vector<double> & its4Chi2, HcalTimeSlew::BiasSetting slewFlavor, int iFitTimes, bool iDCConstraint=false);
 
@@ -126,6 +46,9 @@ public:
     void resetPulseShapeTemplate(const HcalPulseShapes::Shape& ps, unsigned nSamples);
 
 private:
+
+    double getSiPMDarkCurrent(double darkCurrent, double fcByPE, double lambda) const;
+
     int pulseShapeFit(const double * energyArr, const double * pedenArr, const double *chargeArr, 
 		      const double *pedArr, const double *gainArr, const double tsTOTen, std::vector<float> &fitParsVec, const double * ADCnoise, unsigned int soi) const;
     void fit(int iFit,float &timevalfit,float &chargevalfit,float &pedvalfit,float &chi2,bool &fitStatus,double &iTSMax,
