+++ RecoLocalCalo/HcalRecAlgos/src/HcalSimpleRecAlgo.cc
@@ -50,52 +50,7 @@ void HcalSimpleRecAlgo::setRecoParams(bool correctForTimeslew, bool correctForPu
    pileupCleaningID_=pileupCleaningID;
 }
 
-void HcalSimpleRecAlgo::setpuCorrParams(bool   iPedestalConstraint, bool iTimeConstraint,bool iAddPulseJitter,
-					bool iApplyTimeSlew,double iTS4Min, const std::vector<double> & iTS4Max,
-					double iPulseJitter,
-					double iTimeMean, double iTimeSig, double iTimeSigSiPM,
-					double iPedMean, double iPedSig, double iPedSigSiPM,
-					double iNoise, double iNoiseSiPM,
-					double iTMin,double iTMax,
-					const std::vector<double> & its4Chi2, int iFitTimes) {
-  if( iPedestalConstraint ) assert ( iPedSig );
-  if( iTimeConstraint ) assert( iTimeSig );
-  psFitOOTpuCorr_->setPUParams(iPedestalConstraint,iTimeConstraint,iAddPulseJitter,iApplyTimeSlew,
-			       iTS4Min, iTS4Max, iPulseJitter,iTimeMean,iTimeSig,iTimeSigSiPM,iPedMean,iPedSig,iPedSigSiPM,iNoise,iNoiseSiPM,iTMin,iTMax,its4Chi2,
-			       HcalTimeSlew::Medium, iFitTimes);
-
-  psFitOOTpuCorr_->setChi2Term(true); // isHPD all the time
-
-//  int shapeNum = HPDShapev3MCNum;
-//  psFitOOTpuCorr_->setPulseShapeTemplate(theHcalPulseShapes_.getShape(shapeNum));
-}
-
-void HcalSimpleRecAlgo::setMeth3Params( bool iApplyTimeSlew, float iPedSubThreshold, int iTimeSlewParsType, std::vector<double> iTimeSlewPars, double irespCorrM3) {
-
-  pedSubFxn_->init(0, iPedSubThreshold, 0.0);
-  hltOOTpuCorr_->init((HcalTimeSlew::ParaSource)iTimeSlewParsType, HcalTimeSlew::Medium, iApplyTimeSlew, *pedSubFxn_, iTimeSlewPars,irespCorrM3);
-
-}
-
-void HcalSimpleRecAlgo::setForData (int runnum) { 
-   runnum_ = runnum;
-   if( puCorrMethod_ ==2 ){
-      int shapeNum = HPDShapev3MCNum;
-      if( runnum_ > 0 ){
-         shapeNum = HPDShapev3DataNum;
-      }
-      bool isHPD=true;
-      psFitOOTpuCorr_->setPulseShapeTemplate(theHcalPulseShapes_.getShape(shapeNum),isHPD,CaloSamples::MAXSAMPLES);
-   }
-}
-
 void HcalSimpleRecAlgo::setLeakCorrection () { setLeakCorrection_ = true;}
-
-void HcalSimpleRecAlgo::setHBHEPileupCorrection(
-     boost::shared_ptr<AbsOOTPileupCorrection> corr)
-{
-    hbhePileupCorr_ = corr;
-}
 
 void HcalSimpleRecAlgo::setHFPileupCorrection(
      boost::shared_ptr<AbsOOTPileupCorrection> corr)
@@ -323,7 +278,6 @@ namespace HcalSimpleRecAlgoImpl {
     bool leakCorrApplied = false;
     float t0 =0, t2 =0;
     float time = -9999;
-    float m3_time = -9999;
 
 // Disable method 1 inside the removePileup function this way!
 // Some code in removePileup does NOT do pileup correction & to make sure maximum share of code
@@ -352,49 +306,6 @@ namespace HcalSimpleRecAlgoImpl {
       time=time-calibs.timecorr(); // time calibration
     }
 
-    // Note that uncorr_ampl is always set from outside of method 2!
-    if( puCorrMethod == 2 ){
-
-      bool useTriple=false;
-      float chi2=-1;
-
-      CaloSamples cs;
-      coder.adc2fC(digi,cs);
-      std::vector<int> capidvec;
-      for(int ip=0; ip<cs.size(); ip++){
-        const int capid = digi[ip].capid();
-        capidvec.push_back(capid);
-      }
-      psFitOOTpuCorr->apply(cs, capidvec, calibs, ampl, time, useTriple,chi2);
-    }
-    
-    // S. Brandt - Feb 19th : Adding Section for HLT
-    // Run "Method 3" all the time.
-    {
-
-      CaloSamples cs;
-      coder.adc2fC(digi,cs);
-      std::vector<int> capidvec;
-      for(int ip=0; ip<cs.size(); ip++){
-        const int capid = digi[ip].capid();
-        capidvec.push_back(capid);
-      }
-      hltOOTpuCorr->apply(cs, capidvec, calibs, digi, m3_ampl,m3_time);
-      if (puCorrMethod == 3) {ampl = m3_ampl; time=m3_time;}
-    }
-
-    // Temporary hack to apply energy-dependent corrections to some HB- cells
-    if (runnum > 0) {
-      const HcalDetId& cell = digi.id();
-      if (cell.subdet() == HcalBarrel) {
-        const int ieta = cell.ieta();
-        const int iphi = cell.iphi();
-        uncorr_ampl *= hbminus_special_ecorr(ieta, iphi, uncorr_ampl, runnum);
-        ampl *= hbminus_special_ecorr(ieta, iphi, ampl, runnum);
-        m3_ampl *= hbminus_special_ecorr(ieta, iphi, m3_ampl, runnum);
-      }
-    }
-
     // Correction for a leak to pre-sample
     if(useLeak && !leakCorrApplied) {
       uncorr_ampl *= leakCorr(uncorr_ampl);
@@ -407,126 +318,7 @@ namespace HcalSimpleRecAlgoImpl {
     setAuxEnergy(rh, static_cast<float>(m3_ampl));
     return rh;
   }
-
-  template<class Digi, class RecHit>
-  inline RecHit recoHBHE(const Digi& digi, const HcalCoder& coder,
-			 const HcalCalibrations& calibs, 
-			 const int ifirst, const int n, const bool slewCorrect,
-			 const bool pulseCorrect, const HcalPulseContainmentCorrection* corr,
-			 const HcalTimeSlew::BiasSetting slewFlavor,
-			 const int runnum, const bool useLeak,
-			 const AbsOOTPileupCorrection* pileupCorrection,
-			 const BunchXParameter* bxInfo, const unsigned lenInfo, const int puCorrMethod, const PulseShapeFitOOTPileupCorrection * psFitOOTpuCorr, HcalDeterministicFit * hltOOTpuCorr, PedestalSub * hltPedSub)// const on end
-  {
-    double fc_ampl =0, ampl =0, uncorr_ampl =0, m3_ampl =0, maxA = -1.e300;
-    int nRead = 0, maxI = -1;
-    bool leakCorrApplied = false;
-    float t0 =0, t2 =0;
-    float time = -9999;
-    bool useTriple = false;
-    float chi2 = -1;
-    
-    // Disable method 1 inside the removePileup function this way!
-    // Some code in removePileup does NOT do pileup correction & to make sure maximum share of code
-    const AbsOOTPileupCorrection * inputAbsOOTpuCorr = ( puCorrMethod == 1 ? pileupCorrection: nullptr );
-    
-    removePileup(digi, coder, calibs, ifirst, n,
-		 pulseCorrect, corr, inputAbsOOTpuCorr,
-		 bxInfo, lenInfo, &maxA, &ampl,
-		 &uncorr_ampl, &fc_ampl, &nRead, &maxI,
-		 &leakCorrApplied, &t0, &t2);
-    
-    if (maxI > 0 && maxI < (nRead - 1))
-      {
-	// Handle negative excursions by moving "zero":
-	float minA=t0;
-	if (maxA<minA) minA=maxA;
-	if (t2<minA)   minA=t2;
-	if (minA<0) { maxA-=minA; t0-=minA; t2-=minA; } // positivizes all samples
-	
-	float wpksamp = (t0 + maxA + t2);
-	if (wpksamp!=0) wpksamp=(maxA + 2.0*t2) / wpksamp; 
-	time = (maxI - digi.presamples())*25.0 + timeshift_ns_hbheho(wpksamp);
-	
-	if (slewCorrect) time-=HcalTimeSlew::delay(std::max(1.0,fc_ampl),slewFlavor);
-	
-	time=time-calibs.timecorr(); // time calibration
-      }
-    
-    // Note that uncorr_ampl is always set from outside of method 2!
-    if( puCorrMethod == 2 ){
-
-      CaloSamples cs;
-      coder.adc2fC(digi,cs);
-      std::vector<int> capidvec;
-      for(int ip=0; ip<cs.size(); ip++){
-	const int capid = digi[ip].capid();
-	capidvec.push_back(capid);
-      }
-      psFitOOTpuCorr->apply(cs, capidvec, calibs, ampl, time, useTriple,chi2);
-    }
-    
-    // S. Brandt - Feb 19th : Adding Section for HLT
-    // Run "Method 3" all the time.
-    {
-
-      CaloSamples cs;
-      coder.adc2fC(digi,cs);
-      std::vector<int> capidvec;
-      for(int ip=0; ip<cs.size(); ip++){
-	const int capid = digi[ip].capid();
-	capidvec.push_back(capid);
-      }
-
-      float m3_time=0;
-
-      hltOOTpuCorr->apply(cs, capidvec, calibs, digi, m3_ampl, m3_time);
-      if (puCorrMethod == 3) { ampl = m3_ampl; time = m3_time; }
-
-    }
-
-    // Temporary hack to apply energy-dependent corrections to some HB- cells
-    if (runnum > 0) {
-      const HcalDetId& cell = digi.id();
-      if (cell.subdet() == HcalBarrel) {
-	const int ieta = cell.ieta();
-	const int iphi = cell.iphi();
-	uncorr_ampl *= hbminus_special_ecorr(ieta, iphi, uncorr_ampl, runnum);
-        ampl *= hbminus_special_ecorr(ieta, iphi, ampl, runnum);
-        m3_ampl *= hbminus_special_ecorr(ieta, iphi, m3_ampl, runnum);
-      }
-    }
-
-    // Correction for a leak to pre-sample
-    if(useLeak && !leakCorrApplied) {
-      uncorr_ampl *= leakCorr(uncorr_ampl); 
-      if (puCorrMethod < 2)
-        ampl *= leakCorr(ampl); 
-    }
-
-    HBHERecHit rh(digi.id(),ampl,time);
-    if(useTriple)
-      {
-	rh.setFlagField(1, HcalCaloFlagLabels::HBHEPulseFitBit);
-      }
-    rh.setChiSquared(chi2);
-    setRawEnergy(rh, static_cast<float>(uncorr_ampl));
-    setAuxEnergy(rh, static_cast<float>(m3_ampl));
-    return rh;
-  }
 }
-
-
-HBHERecHit HcalSimpleRecAlgo::reconstruct(const HBHEDataFrame& digi, int first, int toadd, const HcalCoder& coder, const HcalCalibrations& calibs) const {
-  return HcalSimpleRecAlgoImpl::recoHBHE<HBHEDataFrame,HBHERecHit>(digi,coder,calibs,
-							       first,toadd,correctForTimeslew_, correctForPulse_,
-							       pulseCorr_->get(digi.id(), toadd, phaseNS_),
-							       HcalTimeSlew::Medium,
-                                                               runnum_, setLeakCorrection_,
-                                                               hbhePileupCorr_.get(),
-                                                               bunchCrossingInfo_, lenBunchCrossingInfo_, puCorrMethod_, psFitOOTpuCorr_.get(),/*hlt*/hltOOTpuCorr_.get(),pedSubFxn_.get());
-}
-
 
 HORecHit HcalSimpleRecAlgo::reconstruct(const HODataFrame& digi, int first, int toadd, const HcalCoder& coder, const HcalCalibrations& calibs) const {
   return HcalSimpleRecAlgoImpl::reco<HODataFrame,HORecHit>(digi,coder,calibs,
