+++ RecoLocalCalo/HcalRecProducers/src/HcalSimpleReconstructor.cc
@@ -25,17 +25,8 @@ HcalSimpleReconstructor::HcalSimpleReconstructor(edm::ParameterSet const& conf):
   paramTS(nullptr),
   theTopology(nullptr)
 {
-  // Intitialize "method 3"
-  reco_.setMeth3Params(
-	    conf.getParameter<bool>    ("applyTimeSlewM3"),
-            conf.getParameter<double>  ("pedestalUpperLimit"),
-            conf.getParameter<int>     ("timeSlewParsType"),
-            conf.getParameter<std::vector<double> >("timeSlewPars"),
-            conf.getParameter<double>  ("respCorrM3")
-      );
 
   // register for data access
-  tok_hbhe_ = consumes<HBHEDigiCollection>(inputLabel_);
   tok_hf_ = consumes<HFDigiCollection>(inputLabel_);
   tok_ho_ = consumes<HODigiCollection>(inputLabel_);
   tok_calib_ = consumes<HcalCalibDigiCollection>(inputLabel_);
@@ -45,10 +36,6 @@ HcalSimpleReconstructor::HcalSimpleReconstructor(edm::ParameterSet const& conf):
     subdet_=HcalOuter;
     produces<HORecHitCollection>();
   }  
-  else if (!strcasecmp(subd.c_str(),"HBHE")) {
-    subdet_=HcalBarrel;
-    produces<HBHERecHitCollection>();
-  } 
   else if (!strcasecmp(subd.c_str(),"HF")) {
     subdet_=HcalForward;
     produces<HFRecHitCollection>();
@@ -57,17 +44,6 @@ HcalSimpleReconstructor::HcalSimpleReconstructor(edm::ParameterSet const& conf):
     std::cout << "HcalSimpleReconstructor is not associated with a specific subdetector!" << std::endl;
   }       
   
-}
-
-void HcalSimpleReconstructor::fillDescriptions(edm::ConfigurationDescriptions& descriptions) {
-  edm::ParameterSetDescription desc;
-  desc.setAllowAnything();
-  desc.add<bool>("applyTimeSlewM3", true);
-  desc.add<double>("pedestalUpperLimit", 2.7); 
-  desc.add<int>("timeSlewParsType",3);
-  desc.add<std::vector<double>>("timeSlewPars", { 12.2999, -2.19142, 0, 12.2999, -2.19142, 0, 12.2999, -2.19142, 0 });
-  desc.add<double>("respCorrM3", 1.0);
-  descriptions.add("simpleHbhereco",desc);
 }
 
 HcalSimpleReconstructor::~HcalSimpleReconstructor() { 
@@ -142,13 +118,8 @@ void HcalSimpleReconstructor::process(edm::Event& e, const edm::EventSetup& even
 
 void HcalSimpleReconstructor::produce(edm::Event& e, const edm::EventSetup& eventSetup)
 {
-  // HACK related to HB- corrections
-  if(e.isRealData()) reco_.setForData(e.run());
- 
   if (det_==DetId::Hcal) {
-    if ((subdet_==HcalBarrel || subdet_==HcalEndcap)) {
-      process<HBHEDigiCollection, HBHERecHitCollection>(e, eventSetup, tok_hbhe_);
-    } else if (subdet_==HcalForward) {
+    if (subdet_==HcalForward) {
       process<HFDigiCollection, HFRecHitCollection>(e, eventSetup, tok_hf_);
     } else if (subdet_==HcalOuter) {
       process<HODigiCollection, HORecHitCollection>(e, eventSetup, tok_ho_);
